Refactor person and cargo models: Change Cargo ID type to INTEGER, update Persona ID to UUID with default value, and enhance createPerson function to validate cargo existence and include cargo details in responses.

+++ src/models/Cargo.js
@@ -3,7 +3,7 @@ import sequelize from '../config/database.js';
 
 const Cargo = sequelize.define('Cargo', {
   id: {
-    type: DataTypes.NUMBER(3),
+    type: DataTypes.INTEGER,
     primaryKey: true,
   },
   nombre: {
